tree-wide: consolidate reboot handling
This patch tries to consolidate the reboot handling of SWUpdate.

Currently, there are multiple places a reboot can be configured. Some use
cases cannot be implemented with the current implementation.

If one wants to differentiate per update if a reboot of the target is
needed, this must be done with the swupdate-progress tool (or any other
mechanism listening to the IPC socket). This eliminates the possibility
to restart the target via the web interface (mongoose). The web
interface triggers the post-update script if one is defined.

Defining the post-update eliminates the use case where an update doesn't
require a reboot of the target, because the post-update command is
executed in any case.

The patch tries to move the reboot handling to one place.

This change has only been roughly tested so far.
I would like to receive some feedback if this something you could
imagine to merge in the future.

If I missed some use cases (or anything else) - please let me know.

Signed-off-by: Daniel Braunwarth <oss@braunwarth.dev>